+++ NEWS.md
@@ -1,3 +1,27 @@
+## meta, version 6.2-0 (2023-mm-dd)
+
+### Major changes
+
+* New function trimfill.rm5() to conduct trim-and-fill method for all
+  or selected meta-analyses of a Cochrane review
+
+### User-visible changes
+
+* nnt.meta(), print.nnt.meta():
+  - NNTs for hazard ratios can be calculated following [Altman &
+    Andersen (1999)](https://doi.org/10.1136/bmj.319.7223.1492)
+  - print sensible confidence interval for NNTs if treatment effect is
+    non-significant [(Altman,
+    1998)](https://doi.org/10.1136/bmj.317.7168.1309)
+
+### Internal changes
+
+* Use generic functions for cumulative meta-analysis (metacum),
+  leave-one-out meta-analysis (metainf) and meta-regression (metareg)
+
+* New internal function chksuitable() to check for suitable classes
+
+
 ## meta, version 6.1-0 (2022-12-20)
 
 ### Major changes